+++ msvsphere/rpmqc/runner.py
@@ -24,10 +24,10 @@ def load_inspections(cfg: Config) -> List[PkgBaseInspector]:
     """
     # TODO: use dynamical loading via pkgutil
     inspections = []
-    from .inspectors.pkg_signature_inspector import PkgSignatureInspectorInspector
-    from .inspectors.pkg_tags_inspector import PkgTagsInspectorInspector
-    inspections.append(PkgSignatureInspectorInspector(cfg))
-    inspections.append(PkgTagsInspectorInspector(cfg))
+    from .inspectors.pkg_signature_inspector import PkgSignatureInspector
+    from .inspectors.pkg_tags_inspector import PkgTagsInspector
+    inspections.append(PkgSignatureInspector(cfg))
+    inspections.append(PkgTagsInspector(cfg))
     return inspections